backend/pyproject.toml
[tool.poetry]
name = "medical-transcription-backend"
version = "0.1.0"
description = "Production-grade FastAPI backend for medical transcription"
authors = ["Your Team <team@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
SQLAlchemy = "^2.0"
pydantic-settings = "^2.1"
python-jose = {extras = ["cryptography"], version = "^3.3"}
passlib = {extras = ["bcrypt"], version = "^1.7"}
python-multipart = "^0.0.6"
redis = "^5.0"
rq = "^1.15"
boto3 = "^1.34"
prometheus-client = "^0.20.0"
alembic = "^1.13"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.23"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"

